+++ package.json
@@ -40,6 +40,7 @@
     "leaflet.markercluster": "^0.4.0",
     "lodash": "^3.10.1",
     "react": "^0.12.2",
+    "react-router": "^0.11.5",
     "whatwg-fetch": "^0.10.0"
   },
   "browserify": {
